Enhance backend services: update course offering, enrollment, program, and student modules; add room management (controller, service, repository, model, DTOs); introduce bulk enrollment/course offering requests and scheduling utilities; remove old GlobalExceptionHandler

+++ src/main/java/com/hopesapms/app/dto/CourseOfferingRequestDTO.java
@@ -18,7 +18,7 @@ public class CourseOfferingRequestDTO {
     private Long academicSemesterId;
 
     @NotNull
-    private Integer instructorId;
+    private Long instructorId;
 
     @NotNull
     @Size(min = 1, message = "At least one year level must be selected")
